Skeleton of the program defined
Figured out how exactly I want the CLI to work which also defined how
the system as a whole is structured. It is built around 2 independent
variables: the data type being mined, and the mining algorithm. The data
type does not influence the miner and the miner does not influence the
data type

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>