Rewrote how itemsets are stored.
Now instead of the massive duplication that was resulting in how i was
storing parents and children of the itemsets nodes in the cache it has
been reduced. The transactions for each itemset is now only stored in
the embeddings list. The parents and children trees now store just what
items are in the itemsets instead of also storing the embeddings. It
makes much more sense and is more efficient.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>