Itemset indices are now fully in memory
The idea that we would store them in the B+Tree was cute but ultimately
it is faster to just store them in the array. None of the datasets I am
working with are larger than main memory on my laptop so they fit there.
They are also more compact etc...

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>