refactor Node interface to internalize dt ref.
The way the data types got implemented tied them fairly strongly to the
given support values. This meant that they would not generalize very
well to calls with different supports. I decided to formalize that and
also took the opportunity to move the DataType pointer into the *Node's.
This was initially not done so these Node's would not have to bloat
their instances with the extra pointer but since once again it was
nearly always needed it made the code messy and bloated the stack. So
moving it inside of the *Node seems to simplify things a bit.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>