only finding max in uniprox (will revert)
This commit only finds the maximal items with uniprox. It is not worth
it to do this. It only adds complexity and according to quick-model is
less uniform on its output. It is also much slower because of the
restarts.

I think in the future this could be a mode. But IMO probably better to
not do this.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>