graple can now report probabilites with reporters
This took some major work and the option to turn computation on and off
at graple will probably go away but for right now it stays. The
important thing is that the probabilities can either be:

1. Logged (using the log reporter)
2. Put in one file (using the file reporter)
3. Put in a file per dir (using the dir reporter)

Which is pretty nice it isn't perfect for sure but it is much better
than what was there before. Now other modes can also implement
probability computation and add their own formatter so that the
reporters can report them.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>