graple pr. comp. with python + scipy + umfpack
Doing the probability computations with scipy and umfpack is orders of
magnatude faster than using go. The reason is that umfpack is a highly
optimized sparse matrix implementation that makes the inversion really,
really fast.

The way it works:

1. compute the LU decomposition of Q
2. in the case where Q is small (< 30000) rows solve Q x = I with
   lu.solve(Q, I) where I is the identity matrix
3. in the case where Q is large solve Q x = I column by column with
   lu.solve.

This solution is faster than:

1. just using a regular inversion function (which just falls back to
   doing an N^2 algorithm).
2. using a dense matrix
3. using scipy's built in LU solver
4. using scipy's built in sparse solver

In conclusion: SuiteSparse (where umfpack comes from) rocks as long as
you use column by column solving for very large matrices.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>