rewrote subgraph/SubGraph.Embeddings() to iterable
I rewrote the Embeddings() search function so that it operated as a
depth first search. I then made it iterable so that I could have a
client that only consumed some of the embeddings. Should work well for
that purpose. The public API of SubGraph got two new method
DoEmbeddings() and IterEmbeddings().

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>