--unsup-emb-pruning now based on single vertices
This is a very disappointing optimization. I am not sure that it will
ever pay off. It is definitely doing some pruning but either the pruning
is not at all helpful or the overhead of running the pruning
(maintaining the pruning lists/maps) is just too high. Hard to say. It
is also very hard to reason about which vertices are prunable.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>