emb-search-starting-point parameterizes search start strat
The search starting point strategy can make a big difference on the
performance of the system. However, it isn't clear to me what the best
general choice is. It is clear the for the datasets I have tested you
can always do better than random-start but you can also do worse!
However, I don't know if there is universal answer or a master method to
select this parameter. A safe answer is always random-start.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>